+++ api/src/services/users.ts
@@ -9,9 +9,9 @@ import { cloneDeep, isEmpty, mergeWith } from 'lodash-es';
 import { performance } from 'perf_hooks';
 import getDatabase from '../database/index.js';
 import { useLogger } from '../logger.js';
-import { getRoleCountsByUsers } from '../telemetry/utils/check-increased-user-limits.js';
+import { checkIncreasedUserLimits } from '../telemetry/utils/check-increased-user-limits.js';
 import { getRoleCountsByRoles } from '../telemetry/utils/get-role-counts-by-roles.js';
-import { checkIncreasedUserLimits } from '../telemetry/utils/get-role-counts-by-users.js';
+import { getRoleCountsByUsers } from '../telemetry/utils/get-role-counts-by-users.js';
 import { type UserCount } from '../telemetry/utils/get-user-count.js';
 import type { AbstractServiceOptions, MutationOptions } from '../types/index.js';
 import { getSecret } from '../utils/get-secret.js';
